services/streaming/requirements.txt
fastapi
uvicorn
httpx
pydantic-settings
# Dépendances optionnelles pour les pipelines externes
redis>=4.6.0
nats-py>=2.3.0
